+++ src/MMALSharp/Config/SensorDefs/SensorDef.cs
@@ -11,27 +11,27 @@ namespace MMALSharp.Config.SensorDefs
         public List<SensorReg> StopReg { get; set; }
         public int NumStopRegs { get; set; }
 
-        public int I2CAddr { get; set; }
+        public byte I2CAddr { get; set; }
         public int I2CAddressing { get; set; }
         public int I2CDataSize { get; set; }
 
-        public int I2CIdentLength { get; set; }
-        public int I2CIdentReg { get; set; }
-        public int I2CIdentValue { get; set; }
+        public ushort I2CIdentLength { get; set; }
+        public ushort I2CIdentReg { get; set; }
+        public ushort I2CIdentValue { get; set; }
 
-        public int VFlipReg { get; set; }
+        public ushort VFlipReg { get; set; }
         public int VFlipRegBit { get; set; }
-        public int HFlipReg { get; set; }
+        public ushort HFlipReg { get; set; }
         public int HFlipRegBit { get; set; }
         public int FlipsDontChangeBayerOrder { get; set; }
 
-        public int ExposureReg { get; set; }
+        public ushort ExposureReg { get; set; }
         public int ExposureRegNumBits { get; set; }
 
-        public int VtsReg { get; set; }
+        public ushort VtsReg { get; set; }
         public int VtsRegNumBits { get; set; }
 
-        public int GainReg { get; set; }
+        public ushort GainReg { get; set; }
         public int GainRegNumBits { get; set; }
     }
 }
